Use request caching.
This time we're using the database, not memory, so memory leaks shouldn't be an issue.

I'll need to add a cron job to clear the cache occasionally.
Either that or clear it when new projects are uploaded, in which case we could be a little more granular about which pages can stay cached.

Fixes https://github.com/jasononeil/haxelib/issues/24
Fixes https://github.com/jasononeil/haxelib/issues/13